app/requirements.txt
geopy
folium
plotly
cufflinks
geocoder
streamlit_lottie
streamlit_folium